requirements.txt
pyserial==3.4
pyvjoy==1.0.1
